+++ src/modbus-flex-sequencer.js
@@ -201,16 +201,19 @@ module.exports = function (RED) {
       /* istanbul ignore next */
       if (mbBasics.invalidPayloadIn(msg)) {
         verboseWarn('Invalid message on input.')
+        mbBasics.sendEmptyMsgOnFail(node, new Error('Invalid message on input.'), msg)
         return
       }
       /* istanbul ignore next */
       if (node.isNotReadyForInput()) {
         verboseWarn('Inject while node is not ready for input.')
+        mbBasics.sendEmptyMsgOnFail(node, new Error('Inject while node is not ready for input.'), msg)
         return
       }
       /* istanbul ignore next */
       if (modbusClient.isInactive()) {
         verboseWarn('You sent an input to inactive client. Please use initial delay on start or send data more slowly.')
+        mbBasics.sendEmptyMsgOnFail(node, new Error('You sent an input to inactive client. Please use initial delay on start or send data more slowly.'), msg)
         return
       }